move TextControl to common component
