+++ wordpress/wp-content/plugins/cds-web-blocks/resources/js/src/product/edit.js
@@ -1,5 +1,5 @@
 import { useBlockProps } from '@wordpress/block-editor';
-import TextControl from './components/TextControl';
+import TextControl from '../components/TextControl';
 import { __ } from '@wordpress/i18n';
 
 // http://localhost/cds/wp-json/wp/v2/product?_embed
